requirements.txt
selenium==4.15.2
Pillow==10.1.0
requests==2.31.0
openai==1.3.7
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
webdriver-manager==4.0.1
streamlit==1.36.0 